Update user migration to use float instead of integer for postgres compatibility

+++ data/migrations/20191021201102_users.js
@@ -24,7 +24,7 @@ exports.up = knex => {
         .unique()
         .notNullable();
       // table.string('name');
-      table.integer('cooldown');
+      table.float('cooldown');
       // table.integer('encumbrance');
       // table.integer('strength');
       // table.integer('speed');
